+++ src/test/feature-instrinsics.spec.ts
@@ -1,4 +1,4 @@
-import { Address, toNano } from 'ton-core';
+import { Address, beginCell, Cell, toNano } from 'ton-core';
 import { ContractSystem } from 'ton-emulator';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { IntrinsicsTester } from './features/output/intrinsics_IntrinsicsTester';
@@ -19,5 +19,7 @@ describe('feature-instrinsics', () => {
         expect(await contract.getGetString2()).toBe('Hello world');
         expect((await contract.getGetAddress()).equals(Address.parse('EQCD39VS5jcptHL8vMjEXrzGaRcCVYto7HUn4bpAOg8xqB2N'))).toBe(true);
         expect((await contract.getGetAddress2()).equals(Address.parse('EQCD39VS5jcptHL8vMjEXrzGaRcCVYto7HUn4bpAOg8xqB2N'))).toBe(true);
+        expect((await contract.getGetCell()).equals(Cell.fromBase64('te6cckEBAQEADgAAGEhlbGxvIHdvcmxkIXgtxbw='))).toBe(true);
+        expect((await contract.getGetCell2()).equals(Cell.fromBase64('te6cckEBAQEADgAAGEhlbGxvIHdvcmxkIXgtxbw='))).toBe(true);
     });
 });